Revert "[Feature] Update SMBJ."
To work around regression in SMB anonymous authentication.

See also https://github.com/hierynomus/smbj/issues/792

This reverts commit d9ce4e604bd11ff32794966798c3addd13f6293f.

Bug: #1181

+++ app/src/main/java/me/zhanghai/android/files/provider/smb/client/FileByteChannel.kt
@@ -62,7 +62,7 @@ class FileByteChannel(
     override fun onWrite(position: Long, source: ByteBuffer) {
         val sourcePosition = source.position()
         val bytesWritten = try {
-            file.write(ByteBufferChunkProvider(source, position)).toInt()
+            file.write(ByteBufferChunkProvider(source, position))
         } catch (e: SMBRuntimeException) {
             throw e.toIOException()
         }
